requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
httpx
sentence-transformers